Split up dockerprovision connector into files.
